abena_comprehensive_test_suite/Abena Clinical Outcome Framework/requirements.txt
python-dotenv==1.0.0
fastapi==0.110.2
uvicorn==0.29.0
httpx==0.27.0
redis==5.0.1
python-jose[cryptography]==3.3.0 